+++ include/stomptalk/hook_base.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <string_view>
 #include <cassert>
 #include <cstdint>
 
@@ -13,12 +12,12 @@ struct hook_base
     virtual void on_frame(parser_hook&, const char *frame_start) noexcept = 0;
  
     virtual void on_method(parser_hook&, 
-        std::uint64_t method_id, std::string_view) noexcept = 0;
+        std::uint64_t method_id, const char*, std::size_t) noexcept = 0;
  
     virtual void on_hdr_key(parser_hook&, 
-        std::uint64_t hader_id, std::string_view) noexcept = 0;
+        std::uint64_t hader_id, const char*, std::size_t) noexcept = 0;
  
-    virtual void on_hdr_val(parser_hook&, std::string_view) noexcept = 0;
+    virtual void on_hdr_val(parser_hook&, const char*, std::size_t) noexcept = 0;
  
     virtual void on_body(parser_hook&, const void*, std::size_t) noexcept = 0;
